#907: add yarn and corepack, introduce NpmRepository
# Conflicts:
#	cli/src/main/java/com/devonfw/tools/ide/tool/npm/NpmBasedCommandlet.java
#	cli/src/main/java/com/devonfw/tools/ide/tool/repository/ArtifactToolRepository.java
#	cli/src/test/java/com/devonfw/tools/ide/npm/NpmJsonObjectTest.java
