+++ cli/src/main/java/com/devonfw/tools/ide/tool/repository/AbstractToolRepository.java
@@ -3,6 +3,7 @@ package com.devonfw.tools.ide.tool.repository;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -17,6 +18,11 @@ import com.devonfw.tools.ide.tool.ToolCommandlet;
 import com.devonfw.tools.ide.url.model.file.UrlChecksums;
 import com.devonfw.tools.ide.url.model.file.UrlDownloadFileMetadata;
 import com.devonfw.tools.ide.url.model.file.UrlGenericChecksum;
+import com.devonfw.tools.ide.url.model.file.json.ToolDependencies;
+import com.devonfw.tools.ide.url.model.file.json.ToolDependency;
+import com.devonfw.tools.ide.url.model.file.json.ToolSecurity;
+import com.devonfw.tools.ide.url.model.folder.UrlEdition;
+import com.devonfw.tools.ide.url.model.folder.UrlTool;
 import com.devonfw.tools.ide.util.FilenameUtil;
 import com.devonfw.tools.ide.version.GenericVersionRange;
 import com.devonfw.tools.ide.version.VersionIdentifier;
@@ -281,4 +287,34 @@ public abstract class AbstractToolRepository implements ToolRepository {
     List<VersionIdentifier> versions = getSortedVersions(tool, edition, toolCommandlet);
     return VersionIdentifier.resolveVersionPattern(version, versions, this.context);
   }
+
+  @Override
+  public Collection<ToolDependency> findDependencies(String tool, String edition, VersionIdentifier version) {
+
+    UrlEdition urlEdition = this.context.getUrls().getEdition(tool, edition);
+    ToolDependencies dependencies = urlEdition.getDependencyFile().getDependencies();
+    if (dependencies == ToolDependencies.getEmpty()) {
+      UrlTool urlTool = urlEdition.getParent();
+      dependencies = urlTool.getDependencyFile().getDependencies();
+    }
+    if (dependencies != ToolDependencies.getEmpty()) {
+      this.context.trace("Found dependencies in {}", dependencies);
+    }
+    return dependencies.findDependencies(version, this.context);
+  }
+
+  @Override
+  public ToolSecurity findSecurity(String tool, String edition) {
+    UrlEdition urlEdition = this.context.getUrls().getEdition(tool, edition);
+    ToolSecurity security = urlEdition.getSecurityFile().getSecurity();
+    if (security == ToolSecurity.getEmpty()) {
+      UrlTool urlTool = urlEdition.getParent();
+      security = urlTool.getSecurityFile().getSecurity();
+    }
+    if (security != ToolSecurity.getEmpty()) {
+      this.context.trace("Found dependencies in {}", security);
+    }
+    return security;
+  }
+
 }
