+++ cli/src/main/java/com/devonfw/tools/ide/tool/repository/DefaultToolRepository.java
@@ -1,6 +1,5 @@
 package com.devonfw.tools.ide.tool.repository;
 
-import java.util.Collection;
 import java.util.List;
 
 import com.devonfw.tools.ide.context.IdeContext;
@@ -8,11 +7,6 @@ import com.devonfw.tools.ide.os.SystemInfo;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 import com.devonfw.tools.ide.url.model.UrlMetadata;
 import com.devonfw.tools.ide.url.model.file.UrlDownloadFileMetadata;
-import com.devonfw.tools.ide.url.model.file.json.ToolDependencies;
-import com.devonfw.tools.ide.url.model.file.json.ToolDependency;
-import com.devonfw.tools.ide.url.model.file.json.ToolSecurity;
-import com.devonfw.tools.ide.url.model.folder.UrlEdition;
-import com.devonfw.tools.ide.url.model.folder.UrlTool;
 import com.devonfw.tools.ide.url.model.folder.UrlVersion;
 import com.devonfw.tools.ide.version.VersionIdentifier;
 
@@ -44,35 +38,6 @@ public class DefaultToolRepository extends AbstractToolRepository {
     UrlVersion urlVersion = metadata.getVersionFolder(tool, edition, version, toolCommandlet);
     SystemInfo sys = this.context.getSystemInfo();
     return urlVersion.getMatchingUrls(sys.getOs(), sys.getArchitecture());
-  }
-
-  @Override
-  public Collection<ToolDependency> findDependencies(String tool, String edition, VersionIdentifier version) {
-
-    UrlEdition urlEdition = this.context.getUrls().getEdition(tool, edition);
-    ToolDependencies dependencies = urlEdition.getDependencyFile().getDependencies();
-    if (dependencies == ToolDependencies.getEmpty()) {
-      UrlTool urlTool = urlEdition.getParent();
-      dependencies = urlTool.getDependencyFile().getDependencies();
-    }
-    if (dependencies != ToolDependencies.getEmpty()) {
-      this.context.trace("Found dependencies in {}", dependencies);
-    }
-    return dependencies.findDependencies(version, this.context);
-  }
-
-  @Override
-  public ToolSecurity findSecurity(String tool, String edition) {
-    UrlEdition urlEdition = this.context.getUrls().getEdition(tool, edition);
-    ToolSecurity security = urlEdition.getSecurityFile().getSecurity();
-    if (security == ToolSecurity.getEmpty()) {
-      UrlTool urlTool = urlEdition.getParent();
-      security = urlTool.getSecurityFile().getSecurity();
-    }
-    if (security != ToolSecurity.getEmpty()) {
-      this.context.trace("Found dependencies in {}", security);
-    }
-    return security;
   }
 
   @Override
