+++ cli/src/test/java/com/devonfw/tools/ide/npm/NpmJsonObjectTest.java
@@ -34,13 +34,15 @@ class NpmJsonObjectTest extends Assertions {
             "1.0.0": {
               "version": "1.0.0",
               "dist": {
-                "tarball": "https://registry.npmjs.org/npm/-/npm-1.0.0.tgz"
+                "tarball": "https://registry.npmjs.org/npm/-/npm-1.0.0.tgz",
+                "shasum": "sha1-of-1.0.0"
               }
             },
             "2.1.0": {
               "version": "2.1.0",
               "dist": {
-                "tarball": "https://registry.npmjs.org/npm/-/npm-2.1.0.tgz"
+                "tarball": "https://registry.npmjs.org/npm/-/npm-2.1.0.tgz",
+                "shasum": "sha1-of-2.1.0"
               }
             }
           }
@@ -61,7 +63,7 @@ class NpmJsonObjectTest extends Assertions {
   }
 
   private void addVersion(NpmJsonVersions versions, String version) {
-    NpmJsonDist dist = new NpmJsonDist("https://registry.npmjs.org/npm/-/npm-" + version + ".tgz", "f783874393588901af1a4824a145fa009f174d9d");
+    NpmJsonDist dist = new NpmJsonDist("https://registry.npmjs.org/npm/-/npm-" + version + ".tgz", "sha1-of-" + version);
     NpmJsonVersion npmVersion = new NpmJsonVersion(version, dist);
     versions.setDetails(version, npmVersion);
   }
