+++ cli/src/test/resources/ide-projects/ng/repository/node/node/default/windows/npm.cmd
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo "npm windows $*"
